+++ cas-compute/src/symbolic/simplify/rules/trigonometry/consts.rs
@@ -8,16 +8,16 @@ use crate::symbolic::{
 use once_cell::sync::Lazy;
 
 /// The number one, wrapped in an [`Expr`].
-pub static ONE: Lazy<Expr> = Lazy::new(|| Expr::Primary(Primary::Integer(int(1))));
+pub static ONE: Lazy<Expr> = Lazy::new(|| Expr::Primary(Primary::i64(int(1))));
 
 /// The number 1/2, wrapped in an [`Expr`].
 pub static ONE_HALF: Lazy<Expr> = Lazy::new(|| make_fraction(
-    Expr::Primary(Primary::Integer(int(1))),
-    Expr::Primary(Primary::Integer(int(2))),
+    Expr::Primary(Primary::i64(int(1))),
+    Expr::Primary(Primary::i64(int(2))),
 ));
 
 /// The number zero, wrapped in an [`Expr`].
-pub static ZERO: Lazy<Expr> = Lazy::new(|| Expr::Primary(Primary::Integer(int(0))));
+pub static ZERO: Lazy<Expr> = Lazy::new(|| Expr::Primary(Primary::i64(int(0))));
 
 /// Relevant fractions of a unit circle.
 ///
@@ -38,87 +38,87 @@ pub mod input {
 
     /// 1/12 = pi/6 rad = 30 deg
     pub static ONE_TWELFTH: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(1))),
-        Expr::Primary(Primary::Integer(int(12))),
+        Expr::Primary(Primary::i64(int(1))),
+        Expr::Primary(Primary::i64(int(12))),
     ));
 
     /// 1/8 = pi/4 rad = 45 deg
     pub static ONE_EIGHTH: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(1))),
-        Expr::Primary(Primary::Integer(int(8))),
+        Expr::Primary(Primary::i64(int(1))),
+        Expr::Primary(Primary::i64(int(8))),
     ));
 
     /// 1/6 = pi/3 rad = 60 deg
     pub static ONE_SIXTH: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(1))),
-        Expr::Primary(Primary::Integer(int(6))),
+        Expr::Primary(Primary::i64(int(1))),
+        Expr::Primary(Primary::i64(int(6))),
     ));
 
     /// 1/4 = pi/2 rad = 90 deg
     pub static ONE_FOURTH: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(1))),
-        Expr::Primary(Primary::Integer(int(4))),
+        Expr::Primary(Primary::i64(int(1))),
+        Expr::Primary(Primary::i64(int(4))),
     ));
 
     /// 1/3 = 2pi/3 rad = 120 deg
     pub static ONE_THIRD: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(1))),
-        Expr::Primary(Primary::Integer(int(3))),
+        Expr::Primary(Primary::i64(int(1))),
+        Expr::Primary(Primary::i64(int(3))),
     ));
 
     /// 3/8 = 3pi/4 rad = 135 deg
     pub static THREE_EIGHTS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(3))),
-        Expr::Primary(Primary::Integer(int(8))),
+        Expr::Primary(Primary::i64(int(3))),
+        Expr::Primary(Primary::i64(int(8))),
     ));
 
     /// 5/12 = 5pi/6 rad = 150 deg
     pub static FIVE_TWELFTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(5))),
-        Expr::Primary(Primary::Integer(int(12))),
+        Expr::Primary(Primary::i64(int(5))),
+        Expr::Primary(Primary::i64(int(12))),
     ));
 
     // lower half
     /// 7/12 = 7pi/6 rad = 210 deg
     pub static SEVEN_TWELFTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(7))),
-        Expr::Primary(Primary::Integer(int(12))),
+        Expr::Primary(Primary::i64(int(7))),
+        Expr::Primary(Primary::i64(int(12))),
     ));
 
     /// 5/8 = 5pi/4 rad = 225 deg
     pub static FIVE_EIGHTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(5))),
-        Expr::Primary(Primary::Integer(int(8))),
+        Expr::Primary(Primary::i64(int(5))),
+        Expr::Primary(Primary::i64(int(8))),
     ));
 
     /// 2/3 = 4pi/3 rad = 240 deg
     pub static TWO_THIRDS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(2))),
-        Expr::Primary(Primary::Integer(int(3))),
+        Expr::Primary(Primary::i64(int(2))),
+        Expr::Primary(Primary::i64(int(3))),
     ));
 
     /// 3/4 = 3pi/2 rad = 270 deg
     pub static THREE_FOURTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(3))),
-        Expr::Primary(Primary::Integer(int(4))),
+        Expr::Primary(Primary::i64(int(3))),
+        Expr::Primary(Primary::i64(int(4))),
     ));
 
     /// 5/6 = 5pi/3 rad = 300 deg
     pub static FIVE_SIXTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(5))),
-        Expr::Primary(Primary::Integer(int(6))),
+        Expr::Primary(Primary::i64(int(5))),
+        Expr::Primary(Primary::i64(int(6))),
     ));
 
     /// 7/8 = 7pi/4 rad = 315 deg
     pub static SEVEN_EIGHTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(7))),
-        Expr::Primary(Primary::Integer(int(8))),
+        Expr::Primary(Primary::i64(int(7))),
+        Expr::Primary(Primary::i64(int(8))),
     ));
 
     /// 11/12 = 11pi/6 rad = 330 deg
     pub static ELEVEN_TWELFTHS: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(11))),
-        Expr::Primary(Primary::Integer(int(12))),
+        Expr::Primary(Primary::i64(int(11))),
+        Expr::Primary(Primary::i64(int(12))),
     ));
 }
 
@@ -131,22 +131,22 @@ pub mod output {
 
     /// sqrt(2)/2
     pub static SQRT_TWO_HALF: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(2))).sqrt(),
-        Expr::Primary(Primary::Integer(int(2))),
+        Expr::Primary(Primary::i64(int(2))).sqrt(),
+        Expr::Primary(Primary::i64(int(2))),
     ));
 
     /// sqrt(3)/2
     pub static SQRT_THREE_HALF: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(3))).sqrt(),
-        Expr::Primary(Primary::Integer(int(2))),
+        Expr::Primary(Primary::i64(int(3))).sqrt(),
+        Expr::Primary(Primary::i64(int(2))),
     ));
 
     /// sqrt(3)/3
     pub static SQRT_THREE_THIRD: Lazy<Expr> = Lazy::new(|| make_fraction(
-        Expr::Primary(Primary::Integer(int(3))).sqrt(),
-        Expr::Primary(Primary::Integer(int(3))),
+        Expr::Primary(Primary::i64(int(3))).sqrt(),
+        Expr::Primary(Primary::i64(int(3))),
     ));
 
     /// sqrt(3)
-    pub static SQRT_THREE: Lazy<Expr> = Lazy::new(|| Expr::Primary(Primary::Integer(int(3))).sqrt());
+    pub static SQRT_THREE: Lazy<Expr> = Lazy::new(|| Expr::Primary(Primary::i64(int(3))).sqrt());
 }
